+++ lib/sequence/account.rb
@@ -36,61 +36,49 @@ module Sequence
     end
 
     class ClientModule < Sequence::ClientModule
-      # Creates a new account in the ledger.
-      # @param [Hash] opts
-      #   Options hash
-      # @option opts [String] id
-      #   Unique identifier. Auto-generated if not specified.
-      # @option opts [Array<String>] key_ids
+      # Create a new account in the ledger.
+      # @param key_ids [Array<String>]
       #   The key IDs used for signing transactions that spend from the account.
-      # @option opts [Integer] quorum
+      # @param id [String]
+      #   Unique identifier. Auto-generated if not specified.
+      # @param quorum [Integer]
       #   The number of keys required to sign transactions that spend from the
       #   account. Defaults to the number of keys provided.
-      # @option opts [Hash] tags
+      # @param tags [Hash]
       #   User-specified key-value data describing the account.
       # @return [Account]
-      def create(opts = {})
-        validate_inclusion_of!(
-          opts,
-          :id,
-          :key_ids,
-          :quorum,
-          :tags,
+      def create(key_ids:, id: nil, quorum: nil, tags: nil)
+        raise ArgumentError, ':key_ids cannot be empty' if key_ids == []
+        Account.new(
+          client.session.request(
+            'create-account',
+            id: id,
+            key_ids: key_ids,
+            quorum: quorum,
+            tags: tags,
+          ),
         )
-        validate_required!(opts, :key_ids)
-        Account.new(client.session.request('create-account', opts))
       end
 
-      # Updates an account's tags.
-      # @param [Hash] opts
-      #   Options hash
-      # @option opts [String] id
+      # Update an account's tags.
+      # @param id [String]
       #   The ID of the account.
-      # @option opts [Hash] tags
+      # @param tags [Hash]
       #   A new set of tags, which will replace the existing tags.
       # @return [void]
-      def update_tags(opts = {})
-        validate_inclusion_of!(opts, :id, :tags)
-        validate_required!(opts, :id)
-        client.session.request('update-account-tags', opts)
+      def update_tags(id:, tags: nil)
+        raise ArgumentError, ':id cannot be blank' if id == ''
+        client.session.request('update-account-tags', id: id, tags: tags)
       end
 
-      # Filters accounts.
-      #
-      # @param [Hash] opts
-      #   Options hash
-      # @option opts [String] filter
+      # Filter accounts.
+      # @param filter [String]
       #   A filter expression.
-      # @option opts [Array<String|Integer>] filter_params
+      # @param filter_params [Array<String|Integer>]
       #   A list of values that will be interpolated into the filter expression.
       # @return [Query]
-      def list(opts = {})
-        validate_inclusion_of!(
-          opts,
-          :filter,
-          :filter_params,
-        )
-        Query.new(client, opts)
+      def list(filter: nil, filter_params: nil)
+        Query.new(client, filter: filter, filter_params: filter_params)
       end
     end