+++ lib/sequence/client_module.rb
@@ -1,13 +1,9 @@
 # frozen_string_literal: true
 
-require_relative './validations'
-
 module Sequence
   # Base class for ledger client components.
   # @private
   class ClientModule
-    include Sequence::Validations
-
     attr_reader :client
 
     def initialize(client)