+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Upgraded minimum required Ruby version to 2.3.
   [Ruby 2.2 is now EOL and has received its final security
   patch](https://www.ruby-lang.org/en/news/2018/03/28/ruby-2-2-10-released/).
+* Switched from options hash parameter method signatures to named parameters.
 
 ## 1.5 (20180316)